@@ -4,23 +4,18 @@ version: 0.6.1
 publish_to: none
 
 environment:
-  sdk: '>=2.10.0-0 <3.0.0'
+  sdk: '>=2.12.0-259.15.beta <3.0.0'
 
 dependencies:
-  http: ^0.12.0
+  http: '>=0.13.0'
   tekartik_common_utils:
     git:
       url: git://github.com/tekartik/common_utils.dart
-      ref: dart2
-    version: '>=0.9.3'
-  web_socket_channel: any
+      ref: 'null_safety'
+    version: '>=0.1.0'
+  web_socket_channel: '>=2.0.0'
 
 dev_dependencies:
-  test: any
-  dev_test: '>=0.13.2-dev.3'
-  build_runner: any
-  process_run: '>=0.10.0+1'
-
-dependency_overrides:
-  tekartik_http_test:
-    path: ../http_test
+  test: '>=0.16.3'
+  dev_test: '>=0.15.0'
+  process_run: '>=0.12.0+1'
