@@ -22,7 +22,7 @@ class _InternetAddress implements InternetAddress {
   final InternetAddressType type;
 
   @override
-  final String address;
+  final String? address;
 
   _InternetAddress(this.type, this.address);
 
@@ -92,7 +92,7 @@ abstract class InternetAddress {
   /// The numeric address of the host. For IPv4 addresses this is using
   /// the dotted-decimal notation. For IPv6 it is using the
   /// hexadecimal representation.
-  String get address;
+  String? get address;
 
   /*
   /// The host used to lookup the address. If there is no host
@@ -196,7 +196,7 @@ abstract class HttpRequest implements Stream<Uint8List> {
   ///
   /// If the size of the request body is not known in advance,
   /// this value is -1.
-  int get contentLength;
+  int? get contentLength;
 
   /// The method, such as 'GET' or 'POST', for the request.
   String get method;
@@ -316,7 +316,7 @@ abstract class HttpResponse implements StreamSink<Uint8List>, StringSink {
   /// Gets and sets the content length of the response. If the size of
   /// the response is not known in advance set the content length to
   /// -1, which is also the default if not set.
-  int contentLength;
+  int? contentLength;
 
   /// Gets and sets the status code. Any integer value is accepted. For
   /// the official HTTP status codes use the fields from
@@ -326,7 +326,7 @@ abstract class HttpResponse implements StreamSink<Uint8List>, StringSink {
   /// The status code must be set before the body is written
   /// to. Setting the status code after writing to the response body or
   /// closing the response will throw a `StateError`.
-  int statusCode;
+  late int statusCode;
 
   /*
   /// Gets and sets the reason phrase. If no reason phrase is explicitly
@@ -416,7 +416,7 @@ abstract class HttpResponse implements StreamSink<Uint8List>, StringSink {
   /// This operation is non-blocking. See [flush] or [done] for how to get any
   /// errors generated by this call.
   @override
-  void write(Object obj);
+  void write(Object? obj);
 }
 
 /// A server that delivers content, such as web pages, using the HTTP protocol.
@@ -673,7 +673,7 @@ abstract class HttpServer implements Stream<HttpRequest> {
   /// Returns the address that the server is listening on. This can be
   /// used to get the actual address used, when the address is fetched by
   /// a lookup from a hostname.
-  InternetAddress get address;
+  InternetAddress? get address;
 
   /*
   /// Sets the timeout, in seconds, for sessions of this [HttpServer].
@@ -962,7 +962,7 @@ abstract class HttpHeaders {
   /// header will only be updated if this field is set
   /// directly. Mutating the returned current value will have no
   /// effect.
-  ContentType contentType;
+  ContentType? contentType;
 
   /*
 
@@ -977,13 +977,13 @@ abstract class HttpHeaders {
   */
   /// Returns the list of values for the header named [name]. If there
   /// is no header with the provided name, [:null:] will be returned.
-  List<String> operator [](String name);
+  List<String>? operator [](String name);
 
   /// Convenience method for the value for a single valued header. If
   /// there is no header with the provided name, [:null:] will be
   /// returned. If the header has more than one value an exception is
   /// thrown.
-  String value(String name);
+  String? value(String name);
 
   /// Adds a header value. The header named [name] will have the value
   /// [value] added to its list of values. Some headers are single
